ApiCaller broken out from tool.py, implemented verification of authentication, expanded checks on credentials, method for sending messages
- Moved ApiCaller out of tool.py file. Strong possibilty it will be used
in other tools. It might need to be moved further out of the to_splunk
tool, but it should be easy to refactor in the future.
- Implemented checks on credential validity and checks on valid
authentication. Improved expcetion to convey the status code and error
messsage in the event of failed authentication.
